# Add schema cache