<!--
Assign reviewers if ready for review.
 -->

## Description

Fixes #2194

This will not have massive performance effect for most users as we expect most people to just use the default schema, which is already cached. But for people who do set `ObjectClasses`, it should speed things up nicely.

## Benchmarks

### Before the change

|                                                                              Method |      Mean |     Error |    StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------------------------------------------------------------------ |----------:|----------:|----------:|--------:|-------:|------:|----------:|
| CreateSchema |  3.897 ms | 0.0313 ms | 0.0261 ms | 78.1250 | 7.8125 |     - |    859 KB |
| OpenRealm | 16.567 ms | 0.2680 ms | 0.2375 ms | 62.5000 |      - |     - |    898 KB |

### After the change

|                                                                              Method |          Mean |       Error |      StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------------------------------------------------ |--------------:|------------:|------------:|-------:|------:|------:|----------:|
| CreateSchema |      2.128 us |   0.0090 us |   0.0075 us | 0.1793 |     - |     - |      2 KB |
| OpenRealm | 12,254.235 us | 230.4506 us | 215.5636 us |      - |     - |     - |     40 KB |

##  TODO

* [x] Changelog entry
* [x] Tests (if applicable)